% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancova_function.R
\name{compare_ancova}
\alias{compare_ancova}
\title{Perform ANCOVA to Compare Models}
\usage{
compare_ancova(models, covariates, data)
}
\arguments{
\item{models}{A list of linear regression models created using \code{run_regression}.}

\item{covariates}{A vector of strings specifying the covariate variables to be tested.}

\item{data}{A data frame containing the variables used in the models.}
}
\value{
A data frame summarizing ANCOVA results, including F-values and p-values for each covariate in the models.
}
\description{
Performs ANCOVA to compare multiple linear regression models, evaluating the effect of covariates.
}
\examples{
data(mtcars)
model1 <- run_regression(mtcars, response = "mpg", predictors = c("wt"))
model2 <- run_regression(mtcars, response = "mpg", predictors = c("wt", "hp"))
compare_ancova(models = list(model1, model2), covariates = c("wt", "hp"), data = mtcars)
}
