% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_function.R
\name{compare_anova}
\alias{compare_anova}
\title{Compare Models Using ANOVA}
\usage{
compare_anova(models)
}
\arguments{
\item{models}{A list of linear regression models created using \code{run_regression}.}
}
\value{
A data frame containing the ANOVA comparison results, including F-values and p-values for each model.
}
\description{
Performs ANOVA to compare multiple linear regression models.
}
\examples{
data(mtcars)
model1 <- run_regression(mtcars, response = "mpg", predictors = c("wt"))
model2 <- run_regression(mtcars, response = "mpg", predictors = c("wt", "hp"))
compare_anova(models = list(model1, model2))

}
