% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_plot_function.R
\name{plot_anova}
\alias{plot_anova}
\title{Visualize ANOVA Results}
\usage{
plot_anova(model, data, response, predictor)
}
\arguments{
\item{model}{A linear regression model created using \code{run_regression}.}

\item{data}{A data frame containing the variables used in the model.}

\item{response}{A string specifying the response variable.}

\item{predictor}{A string specifying the predictor variable.}
}
\value{
A scatterplot with a regression line.
}
\description{
Creates a scatterplot of the response variable against a predictor, with a regression line based on ANOVA results.
}
\examples{
data(mtcars)
model <- run_regression(mtcars, response = "mpg", predictors = c("wt"))
plot_anova(model, data = mtcars, response = "mpg", predictor = "wt")

}
