% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_function.R
\name{run_regression}
\alias{run_regression}
\title{Run Linear Regression}
\usage{
run_regression(data, response, predictors)
}
\arguments{
\item{data}{A data frame containing the variables in the model.}

\item{response}{A string specifying the response variable.}

\item{predictors}{A vector of strings specifying the predictor variables.}
}
\value{
A linear model object.
}
\description{
Runs a linear regression on the input dataset with specified predictor and response variables.
}
\examples{
data(mtcars)
run_regression(mtcars, response = "mpg", predictors = c("wt", "hp"))
}
